Implementation of the sequential test mechanism (db and mock)
