+++ internal/service/mock_test/runner_mock_test.go
@@ -1,0 +1,270 @@
+package mock_test
+
+import (
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"os"
+	"strconv"
+	"testing"
+
+	"github.com/CrunchyData/pg_featureserv/internal/api"
+	"github.com/CrunchyData/pg_featureserv/internal/conf"
+	"github.com/CrunchyData/pg_featureserv/internal/data"
+	"github.com/CrunchyData/pg_featureserv/internal/service"
+	util "github.com/CrunchyData/pg_featureserv/internal/utiltest"
+	log "github.com/sirupsen/logrus"
+)
+
+var hTest util.HttpTesting
+var catalogMock *data.CatalogMock
+
+// ...
+func TestMain(m *testing.M) {
+	initCatMock()
+	os.Exit(m.Run())
+}
+
+// ...
+type MockTests struct {
+	Test *testing.T
+}
+
+// ...
+func TestRunnerHandlerMock(t *testing.T) {
+	// t.Skip("skipping test.")
+
+	// initialisation avant l'execution des tests
+	beforeRun()
+
+	t.Run("GET", func(t *testing.T) {
+		m := MockTests{Test: t}
+		m.TestCollectionItem()
+		m.TestCollectionItemsResponse()
+		m.TestCollectionMissingItemsNotFound()
+		m.TestCollectionItemPropertiesEmpty()
+		m.TestCollectionNotFound()
+		m.TestCollectionResponse()
+		m.TestCollectionsResponse()
+		m.TestFeatureNotFound()
+	})
+	t.Run("GET - Params", func(t *testing.T) {
+		m := MockTests{Test: t}
+		m.TestRoot()
+		m.TestBBox()
+		m.TestBBoxInvalid()
+		m.TestFilterB()
+		m.TestFilterBD()
+		m.TestFilterBDNone()
+		m.TestFilterD()
+		m.TestLimit()
+		m.TestLimitInvalid()
+		m.TestLimitZero()
+		m.TestOffset()
+		m.TestOffsetInvalid()
+		m.TestProperties()
+		m.TestPropertiesEmpty()
+		m.TestPropertiesAll()
+		m.TestQueryParamCase()
+		m.TestSortBy()
+		m.TestSortByAsc()
+		m.TestSortByDesc()
+		m.TestTransformInvalid()
+		m.TestTransformValid()
+	})
+	t.Run("GET - Html", func(t *testing.T) {
+		m := MockTests{Test: t}
+		m.TestHTMLCollection()
+		m.TestHTMLCollections()
+		m.TestHTMLConformance()
+		m.TestHTMLFunction()
+		m.TestHTMLFunctions()
+		m.TestHTMLItem()
+		m.TestHTMLItems()
+		m.TestHTMLRoot()
+	})
+	t.Run("GET - functions", func(t *testing.T) {
+		m := MockTests{Test: t}
+		m.TestFunctionJSON()
+		m.TestFunctionMissingItemsNotFound()
+		m.TestFunctionNotFound()
+		m.TestFunctionsJSON()
+	})
+	// liste de tests sur la suppression des features
+	t.Run("DELETE", func(t *testing.T) {
+		beforeEachRun()
+		m := MockTests{Test: t}
+		m.TestApiContainsDeleteFeature()
+		m.TestDeleteExistingFeature()
+		m.TestDeleteFeatureErrorMalformedFeatureId()
+		m.TestDeleteFeatureErrorUnknownCollection()
+		m.TestDeleteFeatureErrorUnusedQueryParameters()
+		m.TestDeleteUnknownFeature()
+		afterEachRun()
+	})
+	t.Run("PUT", func(t *testing.T) {
+		beforeEachRun()
+		m := MockTests{Test: t}
+		m.TestApiContainsMethodPut()
+		m.TestGetCollectionReplaceSchema()
+		m.TestReplaceFeatureMissingRequiredPropertiesFailure()
+		m.TestReplaceFeatureOnlyGeomFailure()
+		m.TestReplaceFeatureOnlyPropFailure()
+		m.TestReplaceFeatureRequiredPropertiesSuccess()
+		m.TestReplaceFeatureSuccess()
+		afterEachRun()
+	})
+	t.Run("POST", func(t *testing.T) {
+		beforeEachRun()
+		m := MockTests{Test: t}
+		m.TestApiContainsCollectionSchemas()
+		m.TestApiContainsMethodPostFeature()
+		m.TestGetCollectionCreateSchema()
+		m.TestCreateFeature()
+		afterEachRun()
+	})
+	t.Run("UPDATE", func(t *testing.T) {
+		beforeEachRun()
+		m := MockTests{Test: t}
+		m.TestApiContainsMethodPatchFeature()
+		m.TestGetCollectionUpdateSchema()
+		m.TestUpdateFeatureSuccess()
+		m.TestUpdateFeaturePartialSuccess()
+		m.TestUpdateFeatureOnlyGeomSuccess()
+		m.TestUpdateFeatureOnlyPropSuccess()
+		m.TestUpdateFeaturePartialGeomFailure()
+		afterEachRun()
+	})
+
+	// nettoyage après execution des tests
+	afterRun()
+}
+
+// Run before all tests
+func beforeRun() {
+	log.Debug("beforeRun")
+	// some stuff...
+}
+
+// Run after all tests
+func afterRun() {
+	log.Debug("afterRun")
+	// some stuff...
+}
+
+// Run before each test
+func beforeEachRun() {
+	log.Debug("beforeEachRun")
+	// re init catalog mock...
+	initCatMock()
+}
+
+// Run after each test
+func afterEachRun() {
+	log.Debug("afterEachRun")
+	// some stuff...
+}
+
+// ...
+func initCatMock() {
+	conf.Configuration.Database.AllowWrite = true
+	catalogMock = data.CatMockInstance()
+	service.SetCatalogInstance(catalogMock)
+
+	hTest = util.MakeHttpTesting("http://test", "/pg_featureserv", "../../../assets", service.InitRouter("/pg_featureserv"))
+	service.Initialize()
+}
+
+// ...
+func checkCollection(tb testing.TB, coll *api.CollectionInfo, name string, title string) {
+	util.Equals(tb, name, coll.Name, "Collection name")
+	util.Equals(tb, title, coll.Title, "Collection title")
+
+	path := "/collections/" + name
+	checkLink(tb, coll.Links[0], api.RelSelf, api.ContentTypeJSON, hTest.UrlBase+path)
+	checkLink(tb, coll.Links[1], api.RelAlt, api.ContentTypeHTML, hTest.UrlBase+path+".html")
+
+	pathItems := path + "/items"
+	checkLink(tb, coll.Links[2], api.RelItems, api.ContentTypeGeoJSON, hTest.UrlBase+pathItems)
+}
+
+// ...
+func checkLink(tb testing.TB, link *api.Link, rel string, conType string, href string) {
+	util.Equals(tb, rel, link.Rel, "Link rel")
+	util.Equals(tb, conType, link.Type, "Link type")
+	util.Equals(tb, href, link.Href, "Link href")
+}
+
+// ...
+func checkFunctionSummary(tb testing.TB, v *api.FunctionSummary, fun *api.Function) {
+	util.Equals(tb, fun.Name, v.Name, "Function name")
+	util.Equals(tb, fun.Description, v.Description, "Function description")
+
+	path := "/functions/" + fun.Name
+	checkLink(tb, v.Links[0], api.RelSelf, api.ContentTypeJSON, hTest.UrlBase+path)
+	checkLink(tb, v.Links[1], api.RelAlt, api.ContentTypeHTML, hTest.UrlBase+path+".html")
+
+	pathItems := path + "/items"
+	itemsType := api.ContentTypeJSON
+	if fun.IsGeometryFunction() {
+		itemsType = api.ContentTypeGeoJSON
+	}
+	checkLink(tb, v.Links[2], api.RelItems, itemsType, hTest.UrlBase+pathItems)
+}
+
+// ...
+func checkFunction(t *testing.T, fun *api.Function) {
+	path := "/functions/" + fun.ID
+	resp := hTest.DoRequest(t, path)
+	body, _ := ioutil.ReadAll(resp.Body)
+
+	var v api.FunctionInfo
+	errUnMarsh := json.Unmarshal(body, &v)
+	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
+
+	util.Equals(t, fun.ID, v.Name, "Name")
+	util.Equals(t, fun.Description, v.Description, "Description")
+
+	//--- check parameters
+	util.Assert(t, v.Parameters != nil, "Parameters property must be present")
+	util.Equals(t, len(fun.InNames), len(v.Parameters), "Parameters len")
+	for i := 0; i < len(v.Parameters); i++ {
+		util.Equals(t, fun.InNames[i], v.Parameters[i].Name, "Parameters[].Name")
+		util.Equals(t, fun.InDbTypes[i], v.Parameters[i].Type, "Parameters[].Type")
+	}
+
+	//--- check properties
+	util.Assert(t, v.Properties != nil, "Properties property must be present")
+	util.Equals(t, len(fun.OutNames), len(v.Properties), "Properties len")
+	for i := 0; i < len(v.Properties); i++ {
+		util.Equals(t, fun.OutNames[i], v.Properties[i].Name, "Properties[].Name")
+		util.Equals(t, string(fun.OutJSONTypes[i]), v.Properties[i].Type, "Properties[].Type")
+	}
+
+	//--- check links
+	checkLink(t, v.Links[0], api.RelSelf, api.ContentTypeJSON, hTest.UrlBase+path)
+	checkLink(t, v.Links[1], api.RelAlt, api.ContentTypeHTML, hTest.UrlBase+path+".html")
+	itemsType := api.ContentTypeJSON
+	if fun.IsGeometryFunction() {
+		itemsType = api.ContentTypeGeoJSON
+	}
+	checkLink(t, v.Links[2], api.RelItems, itemsType, hTest.UrlBase+path+"/items")
+}
+
+// check if item is available and is not empty
+func checkItem(t *testing.T, id int) []byte {
+	path := fmt.Sprintf("/collections/mock_a/items/%d", id)
+	resp := hTest.DoRequest(t, path)
+	body, _ := ioutil.ReadAll(resp.Body)
+
+	var v api.GeojsonFeatureData
+	errUnMarsh := json.Unmarshal(body, &v)
+	util.Assert(t, errUnMarsh == nil, fmt.Sprintf("%v", errUnMarsh))
+
+	util.Equals(t, "Feature", v.Type, "feature type")
+	actId, _ := strconv.Atoi(v.ID)
+	util.Equals(t, id, actId, "feature id")
+	util.Equals(t, 4, len(v.Props), "# feature props")
+
+	return body
+}
